feat(resources): extract asset list item component (Phase 6e)
- Extract 'AssetListItem' from the assets list page into a dedicated feature component.
- Refactor the Assets list page to use the new component.
- This completes the component extraction phase for the Resources module.
